Updated README and added comments throughout code
- Updated README with detailed Notes/Assumptions and explanation of frontend total calculation
- Added consistent comments for all backend and frontend files
- Clarified functionality for getAllOrders/getOrders and submitOrder
- Updated database schemas to use DECIMAL for prices

+++ backend/src/models/beverage.ts
@@ -1,13 +1,21 @@
+/**
+ * Beverage interface
+ * Represents a beverage item in the application
+ */
 export interface Beverage {
   id?: number;
   name: string;
   price: number;
 }
 
+/**
+ * SQLite table schema for beverages
+ * Creates table if it doesn't exist
+ */
 export const BeverageTableSchema = `
 CREATE TABLE IF NOT EXISTS beverages (
   id INTEGER PRIMARY KEY,
   name TEXT NOT NULL,
-  price REAL NOT NULL
+  price DECIMAL(10,2) NOT NULL
 )
 `;
