+++ backend/src/models/order.ts
@@ -1,5 +1,9 @@
 import { OrderItem } from "./orderItem";
 
+/**
+ * Order interface
+ * Represents a customer order in the application
+ */
 export interface Order {
   customerName: string;
   customerEmail: string;
@@ -7,11 +11,15 @@ export interface Order {
   total: number;
 }
 
+/**
+ * SQLite table schema for orders
+ * Creates table if it doesn't exist
+ */
 export const OrdersTableSchema = `
 CREATE TABLE IF NOT EXISTS orders (
   id INTEGER PRIMARY KEY,
   customer_name TEXT NOT NULL,
   customer_email TEXT NOT NULL,
-  total REAL NOT NULL
+  total DECIMAL(10,2) NOT NULL
 );
 `
